refactor: prune ecstore root global facades
