+++ crates/ecstore/src/bucket/lifecycle/bucket_lifecycle_ops.rs
@@ -890,7 +890,7 @@ impl TransitionState {
         tokio::spawn(async move {
             Self::inc_counter(&state.compensation_running_tasks);
             state.record_scanner_transition_state();
-            let Some(api) = crate::resolve_object_store_handle() else {
+            let Some(api) = crate::global::resolve_object_store_handle() else {
                 scheduled.lock().unwrap().remove(&bucket);
                 Self::add_counter(&state.compensation_running_tasks, -1);
                 state.record_scanner_transition_state();
@@ -1909,7 +1909,7 @@ pub async fn enqueue_immediate_expiry(oi: &ObjectInfo, src: LcEventSrc) {
     let Some(lifecycle) = GLOBAL_LifecycleSys.get(&oi.bucket).await else {
         return;
     };
-    let Some(api) = crate::resolve_object_store_handle() else {
+    let Some(api) = crate::global::resolve_object_store_handle() else {
         return;
     };
