+++ crates/ecstore/src/lib.rs
@@ -57,13 +57,6 @@ mod pools_test;
 mod store_test;
 mod tier;
 
-pub use global::set_global_endpoints;
-pub use global::update_erasure_type;
-pub use global::{get_global_lock_client, get_global_lock_clients, set_global_lock_client, set_global_lock_clients};
-pub use global::{new_object_layer_fn, resolve_object_store_handle, set_object_store_resolver};
-
-pub use global::GLOBAL_Endpoints;
-
 #[cfg(test)]
 mod rio_tests {
     #[test]
